Now adding attacks works but when we try to get the character, its attacks list is empty

+++ CharacterSheetCreatorBack/CharacterSheetCreatorBack/DAL/CharacterRepository.cs
@@ -21,17 +21,9 @@ namespace CharacterSheetCreatorBack.DAL
 
         public Character? GetCharacter(int idPlayer, int idGame)
         {
-            try
-            {
-                CharacterModel? model =  _rpgContext.Characters.First<CharacterModel>(x =>
-                        x.IdGame == idGame && x.IdPlayer == idPlayer);
-                return model.ToCharacter();
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine(e.ToString());
-                throw new InvalidOperationException("The character doesn't exist.");
-            }
+            CharacterModel? model =  _rpgContext.Characters.First<CharacterModel>(x =>
+                    x.IdGame == idGame && x.IdPlayer == idPlayer);
+            return model.ToCharacter();
         }
 
         /**********************************************************************/
@@ -40,38 +32,35 @@ namespace CharacterSheetCreatorBack.DAL
 
         public int UpdateCharacter(Character character)
         {
-            try
+            // on ne peut pas appeler Update sur le character pris en paramètre
+            // (la méthode Update regarde la référence de l'objet).
+            CharacterModel? dbCharacter = _rpgContext.Characters.First<CharacterModel>(x =>
+                    x.IdGame == character.IdGame && x.IdPlayer == character.IdPlayer);
+
+            if (dbCharacter is null)
             {
-                // on ne peut pas appeler Update sur le character pris en paramètre
-                // (la méthode Update regarde la référence de l'objet).
-                CharacterModel? dbCharacter = _rpgContext.Characters.First<CharacterModel>(x =>
-                        x.IdGame == character.IdGame && x.IdPlayer == character.IdPlayer);
-
-                // suppression de toutes les attacks existantes
-                dbCharacter.Attacks.ForEach(attack => _attackRepository.DeleteAttack(attack.Id));
-                dbCharacter.Attacks.Clear();
-
-                dbCharacter.FromCharacter(character);
-
-                // on recrée les attacks
-                dbCharacter.Attacks.ForEach(attack =>
-                {
-                    _attackRepository.CreateAttack(attack);
-                    dbCharacter.Attacks.Append(attack);
-                });
-
-                // make sure that we have the good id for the output
-                character.ID = dbCharacter.ID;
-
-                // update the db and save changes
-                _rpgContext.Characters.Update(dbCharacter);
-                _rpgContext.SaveChanges();
+                throw new InvalidOperationException("Error: the character doesn't exist.");
             }
-            catch (Exception e)
+
+            // suppression de toutes les attacks existantes
+            dbCharacter.Attacks.ForEach(attack => _attackRepository.DeleteAttack(attack.Id));
+            dbCharacter.Attacks.Clear();
+
+            dbCharacter.FromCharacter(character);
+
+            // on recrée les attacks
+            dbCharacter.Attacks.ForEach(attack =>
             {
-                Console.WriteLine(e.ToString());
-                throw new InvalidOperationException("The player doesn't exist.");
-            }
+                attack.Print();
+                _attackRepository.CreateAttack(attack);
+            });
+
+            // make sure that we have the good id for the output
+            character.ID = dbCharacter.ID;
+
+            // update the db and save changes
+            _rpgContext.Characters.Update(dbCharacter);
+            _rpgContext.SaveChanges();
             return character.ID;
         }
 
@@ -81,25 +70,17 @@ namespace CharacterSheetCreatorBack.DAL
 
         public int CreateCharacter(int IdPlayer,  int IdGame)
         {
-            try
+            Character newCharacter = new Character
             {
-                Character newCharacter = new Character
-                {
-                    IdPlayer = IdPlayer,
-                    IdGame = IdGame
-                };
-                CharacterModel model = new CharacterModel(newCharacter);
+                IdPlayer = IdPlayer,
+                IdGame = IdGame
+            };
+            CharacterModel model = new CharacterModel(newCharacter);
 
-                model.Attacks.ForEach(attack => _attackRepository.CreateAttack(attack));
-                _rpgContext.Characters.Add(new CharacterModel(newCharacter));
-                _rpgContext.SaveChanges();
+            _rpgContext.Characters.Add(new CharacterModel(newCharacter));
+            _rpgContext.SaveChanges();
 
-                return newCharacter.ID;
-            }
-            catch(Exception e)
-            {
-                throw e;
-            }
+            return newCharacter.ID;
         }
 
         /**********************************************************************/
@@ -108,23 +89,16 @@ namespace CharacterSheetCreatorBack.DAL
 
         public void DeleteCharacter(int idPlayer, int idGame)
         {
-            try
-            {
-                CharacterModel? toRemove = _rpgContext.Characters.First<CharacterModel>(x =>
-                        x.IdGame == idPlayer && x.IdPlayer == idGame);
+            CharacterModel? toRemove = _rpgContext.Characters.First<CharacterModel>(x =>
+                    x.IdGame == idPlayer && x.IdPlayer == idGame);
 
-                // on supprime les attacks
-                toRemove.Attacks.ForEach(x => _attackRepository.DeleteAttack(x.Id));
-                toRemove.Attacks.Clear();
+            // on supprime les attacks
+            toRemove.Attacks.ForEach(x => _attackRepository.DeleteAttack(x.Id));
+            toRemove.Attacks.Clear();
 
-                // suppression
-                _rpgContext.Characters.Remove(toRemove);
-                _rpgContext.SaveChanges();
-            }
-            catch (Exception e)
-            {
-                throw new InvalidOperationException("Impossible to delete this character: " + e.Message);
-            }
+            // suppression
+            _rpgContext.Characters.Remove(toRemove);
+            _rpgContext.SaveChanges();
         }
     }
 }
